+++ DecoderFFmpeg.h
@@ -27,7 +27,7 @@ public:
 
 	// Seeks to a 'position' in the stream, in seconds.
 	// Returns the new position in seconds.
-	float Seek( const float position ) override;
+	double Seek( const double position ) override;
 
 private:
 	// Deccodes the next chunk of data into the sample buffer, returning whether any data was decoded.
